refactor: move user/auth functions to their own file
