+++ src/commons.py
@@ -11,7 +11,6 @@ import xml.dom.minidom
 from saxonche import PySaxonProcessor, PyXdmNode
 
 from dynaconf import Dynaconf
-import jinja2
 from fastapi import Request
 
 import toml
@@ -197,47 +196,4 @@ def call_action_hook(req: Request,action:str,app:str,prof:str,rec:str,user:str):
             logging.info(f"no hooks!")
             return None
 
-def allowed(user,app,action,default,prof=None,nr=None):
-    config_app_file = f"{settings.URL_DATA_APPS}/{app}/config.toml"
-    if not os.path.isfile(config_app_file):
-        logging.error(f"config file {config_app_file} doesn't exist")
-        if default == "any":
-            return True
-        return False
-    with open(config_app_file, 'r') as f:
-        config = toml.load(f)
-        mode = default
-        if 'access' in config["app"]:
-            if action in config['app']['access']:
-                mode = config['app']['access'][action]
-        if mode == "owner" and user!=None and prof!=None and nr!=None:
-            record_file = f"{settings.URL_DATA_APPS}/{app}/profiles/{prof}/records/record-{nr}.xml"
-            with open(record_file, 'r') as file:
-                rec = file.read()
-                with PySaxonProcessor(license=False) as proc:
-                    rec = proc.parse_xml(xml_text=rec)
-                    xpproc = proc.new_xpath_processor()
-                    xpproc.set_cwd(os.getcwd())
-                    xpproc.declare_namespace('clariah','http://www.clariah.eu/')
-                    xpproc.declare_namespace('cmd','http://www.clarin.eu/cmd/')
-                    xpproc.set_context(xdm_item=rec)
-                    owner = xpproc.evaluate_single(f"string((/*:CMD/*:Header/*:MdCreator,'{def_user(app)}')[1])").get_string_value()
-                    if owner == user:
-                        return True
-        elif mode == "owner" and user!=None:
-                return True
-        elif mode == "users" and user != None:
-            return True
-        elif mode == "any":
-            return True
-    return False
-    
-def def_user(app):
-    config_app_file = f"{settings.URL_DATA_APPS}/{app}/config.toml"
-    with open(config_app_file, 'r') as f:
-        config = toml.load(f)
-        if 'def_user' in config["app"]:
-            return config["app"]['def_user']
-        elif 'def_user' in settings:
-            return settings.def_user
-        return "server"
+
