+++ src/records.py
@@ -6,7 +6,8 @@ import toml
 from datetime import timezone 
 
 from saxonche import PySaxonProcessor
-from src.commons import settings, convert_toml_to_xml, def_user
+from src.commons import settings, convert_toml_to_xml
+from auth.authentication import def_user
 from src.profiles import prof_xml, prof_json
 
 from weasyprint import HTML
